tests/wpt/tests/html/interaction/focus/focusgroup/tentative/sequential-navigation/directional-input-handler-all-types.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Native directional input handler elements block directional exit</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<meta name="assert" content="Native directional input handler elements block directional navigation when focused.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<!-- Each native directional input handler type: entry works, exit is blocked. -->

<!-- Text input -->
<div id="toolbar-text" focusgroup="toolbar">
  <button tabindex=0 id="btn-text-before">Before</button>
  <input id="text-input" type="text" value="test">
  <button tabindex=0 id="btn-text-after">After</button>
</div>

<!-- Textarea -->
<div id="toolbar-textarea" focusgroup="toolbar">
  <button tabindex=0 id="btn-textarea-before">Before</button>
  <textarea id="textarea">Content</textarea>
  <button tabindex=0 id="btn-textarea-after">After</button>
</div>

<!-- Number input (up/down steps the value) -->
<div id="toolbar-number" focusgroup="toolbar">
  <button tabindex=0 id="btn-number-before">Before</button>
  <input id="number-input" type="number" value="5">
  <button tabindex=0 id="btn-number-after">After</button>
</div>

<!-- Range input (left/right adjusts the value) -->
<div id="toolbar-range" focusgroup="toolbar">
  <button tabindex=0 id="btn-range-before">Before</button>
  <input tabindex=0 id="range-input" type="range" value="50">
  <button tabindex=0 id="btn-range-after">After</button>
</div>

<!-- Select (up/down picks options) -->
<div id="toolbar-select" focusgroup="toolbar">
  <button tabindex=0 id="btn-select-before">Before</button>
  <select id="select-input">
    <option>A</option>
    <option>B</option>
  </select>
  <button tabindex=0 id="btn-select-after">After</button>
</div>

<!-- Contenteditable -->
<div id="toolbar-editable" focusgroup="toolbar">
  <button tabindex=0 id="btn-editable-before">Before</button>
  <div id="editable" contenteditable="true" tabindex="0">Editable</div>
  <button tabindex=0 id="btn-editable-after">After</button>
</div>

<script>
  const testCases = [
    { id: 'text-input', name: 'text input' },
    { id: 'textarea', name: 'textarea' },
    { id: 'number-input', name: 'number input' },
    { id: 'range-input', name: 'range input' },
    { id: 'select-input', name: 'select' },
    { id: 'editable', name: 'contenteditable' },
  ];

  for (const { id, name } of testCases) {
    const element = document.getElementById(id);
    const toolbar = element.closest('[focusgroup]');
    const beforeBtn = toolbar.querySelector('button');

    // Test: Directional navigation TO the conflict element works.
    promise_test(async t => {
      await focusAndWait(beforeBtn);
      assert_equals(document.activeElement, beforeBtn, 'Precondition: before button focused');

      await focusAndSendDirectionalInput(beforeBtn, kRight);
      assert_equals(document.activeElement, element,
                    `Right should navigate TO ${name}`);
    }, `${name}: directional navigation TO element works`);

    // Test: Directional navigation FROM the conflict element is blocked.
    promise_test(async t => {
      await focusAndWait(element);
      assert_equals(document.activeElement, element, `Precondition: ${name} is focused`);

      await assert_directional_input_does_not_move_focus(element);
    }, `${name}: directional navigation FROM element is blocked`);
  }
</script>
